+++ gtkoptions.c
@@ -1582,7 +1582,7 @@ static void OptionsOK( GtkWidget *pw, optionswidget *pow ){
   CHECKUPDATE(pow->pwOutputGWC,fOutputWinPC, "set output winpc %s")
   CHECKUPDATE(pow->pwOutputMWCpst,fOutputMatchPC, "set output matchpc %s")
 
-  if(( n = pow->padjDigits->value ) != fOutputDigits ){
+  if(( n = (int)pow->padjDigits->value ) != fOutputDigits ){
     sprintf(sz, "set output digits %d", n );
     UserCommand(sz); 
   }
@@ -1622,17 +1622,17 @@ static void OptionsOK( GtkWidget *pw, optionswidget *pow ){
   CHECKUPDATE(pow->pwConfStart,fConfirm, "set confirm new %s")
   CHECKUPDATE(pow->pwConfOverwrite,fConfirmSave, "set confirm save %s")
   
-  if(( n = pow->padjCubeAutomatic->value ) != cAutoDoubles){
+  if(( n = (int)pow->padjCubeAutomatic->value ) != cAutoDoubles){
     sprintf(sz, "set automatic doubles %d", n );
     UserCommand(sz); 
   }
 
-  if(( n = pow->padjCubeBeaver->value ) != nBeavers){
+  if(( n = (int)pow->padjCubeBeaver->value ) != nBeavers){
     sprintf(sz, "set beavers %d", n );
     UserCommand(sz); 
   }
   
-  if(( n = pow->padjLength->value ) != nDefaultLength){
+  if(( n = (int)pow->padjLength->value ) != nDefaultLength){
     sprintf(sz, "set matchlength %d", n );
     UserCommand(sz); 
   }
@@ -1692,12 +1692,12 @@ static void OptionsOK( GtkWidget *pw, optionswidget *pow ){
 
   EvalCacheStats( NULL, &cCache, NULL, NULL );
 
-  if((n = pow->padjCache->value) != cCache) {
+  if((n = (int)pow->padjCache->value) != cCache) {
     sprintf(sz, "set cache %d", n );
     UserCommand(sz); 
   }
 
-  if((n = pow->padjDelay->value) != nDelay) {
+  if((n = (int)pow->padjDelay->value) != nDelay) {
     sprintf(sz, "set delay %d", n );
     UserCommand(sz); 
   }
@@ -1728,7 +1728,7 @@ static void OptionsOK( GtkWidget *pw, optionswidget *pow ){
   
   if( pow->fChanged == 1 ) 
   { 
-     n = pow->padjSeed->value;
+     n = (int)pow->padjSeed->value;
      sprintf(sz, "set seed %d", n); 
      UserCommand(sz); 
   }
@@ -1805,7 +1805,7 @@ static void OptionsOK( GtkWidget *pw, optionswidget *pow ){
       && animGUI != ANIMATE_SLIDE )
       UserCommand( "set gui animation slide" );
 
-  if( ( n = pow->padjSpeed->value ) != nGUIAnimSpeed ) {
+  if( ( n = (int)pow->padjSpeed->value ) != nGUIAnimSpeed ) {
       sprintf( sz, "set gui animation speed %d", n );
       UserCommand( sz );
   }
